requirements.txt
torch>=2.1
transformers>=4.40
datasets>=2.18
evaluate>=0.4.1
seqeval>=1.2.2
hydra-core>=1.3
omegaconf>=2.3
accelerate>=0.28
clearml>=1.16
optuna>=3.6
pandas>=2.0
numpy>=1.24
matplotlib>=3.7

